+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - 新增 `en_US`、`ru_RU` 语言文件 #17
 - 新增 `/pland list op` 命令 #15
+- 新增 `/pland set language` 命令 #19
 
 ## [0.6.0] - 2025-03-01